Parameter Make.T

type t

The type of terms.

val atom : ?loc:L.t -> int -> t

Make an atom from an non-zero integer. Positive integers denotes variables, and negative integers denote the negation of the variable corresponding to their absolute value.